+++ src/main/java/edu/illinois/library/cantaloupe/resource/ScaleRestrictedException.java
@@ -1,13 +1,13 @@
 package edu.illinois.library.cantaloupe.resource;
 
 import edu.illinois.library.cantaloupe.http.Status;
-import edu.illinois.library.cantaloupe.resource.ResourceException;
 
 public class ScaleRestrictedException extends ResourceException {
 
-    public ScaleRestrictedException() {
+    public ScaleRestrictedException(double maxScale) {
         super(Status.FORBIDDEN,
-                "Requests for scales in excess of 100% are not allowed.");
+                String.format("Requests for scales in excess of %d%% are not allowed.",
+                        Math.round(maxScale * 100)));
     }
 
 }